// src/types/supabase.ts
export interface Database {
  public: {
    Tables: {
// ...
      users_profile: {
        Row: {
          avatar_url: string | null
          can_create_projects: boolean
          can_create_tasks: boolean
          created_at: string
          email: string
          id: string
          name: string
          role: string
        }
        Insert: {
          avatar_url?: string | null
          can_create_projects?: boolean
          can_create_tasks?: boolean
          created_at?: string
          email: string
          id: string
          name: string
          role?: string
        }
        Update: {
          avatar_url?: string | null
          can_create_projects?: boolean
          can_create_tasks?: boolean
          created_at?: string
          email?: string
          id?: string
          name?: string
          role?: string
        }
        Relationships: [
          {
            foreignKeyName: "users_profile_id_fkey"
            columns: ["id"]
            isOneToOne: true
            referencedRelation: "users"
            referencedColumns: ["id"]
          }
        ]
      }
    }
    Views: {
      [_ in never]: never
    }
    Functions: {
      [_ in never]: never
    }
    Enums: {
      [_ in never]: never
    }
    CompositeTypes: {
      [_ in never]: never
    }
  }
}